Make Android Rokt optional via runtime classpath guards.
Extract Rokt bridge code behind RoktKitAvailability, compile against android-rokt-kit with compileOnly via cordova.gradle, and return a clear error when the kit is not installed.
