+++ plugin/src/android/src/main/java/com/mparticle/cordova/MParticleCordovaPlugin.java
@@ -14,19 +14,6 @@ import com.mparticle.consent.ConsentState;
 import com.mparticle.consent.GDPRConsent;
 import com.mparticle.identity.*;
 import com.mparticle.internal.Logger;
-import com.mparticle.kits.MParticleRoktKt;
-import com.rokt.roktsdk.CacheConfig;
-import com.rokt.roktsdk.RoktConfig;
-import com.rokt.roktsdk.RoktEvent;
-
-import kotlin.Unit;
-import kotlin.coroutines.Continuation;
-import kotlin.jvm.functions.Function2;
-import kotlinx.coroutines.CoroutineScope;
-import kotlinx.coroutines.CoroutineScopeKt;
-import kotlinx.coroutines.Dispatchers;
-import kotlinx.coroutines.flow.Flow;
-import kotlinx.coroutines.flow.FlowKt;
 
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaPlugin;
@@ -44,7 +31,6 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CancellationException;
 
 import static android.R.attr.type;
 
@@ -52,11 +38,8 @@ public class MParticleCordovaPlugin extends CordovaPlugin {
 
     private final static String LOG_TAG = "MParticleCordovaPlugin";
 
-    // Active Rokt event subscriptions, keyed by placement identifier. Each holds the
-    // CoroutineScope collecting that identifier's event Flow so it can be torn down,
-    // preventing leaked scopes, duplicate subscriptions, and delivery through a stale
-    // callbackContext after the WebView is destroyed.
-    private final Map<String, CoroutineScope> roktEventScopes = new HashMap<>();
+    private RoktPluginDelegate roktDelegate;
+    private boolean roktDelegateInitialized;
 
     public boolean execute(String action, final JSONArray args, final CallbackContext callbackContext) throws JSONException {
         if (action.equals("logEvent")) {
@@ -112,32 +95,56 @@ public class MParticleCordovaPlugin extends CordovaPlugin {
             removeCCPAConsentState(args);
             return true;
         } else if (action.equals("selectPlacements")) {
-            selectPlacements(args);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.selectPlacements(args, callbackContext));
         } else if (action.equals("selectShoppableAds")) {
-            selectShoppableAds(args, callbackContext);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.selectShoppableAds(args, callbackContext));
         } else if (action.equals("purchaseFinalized")) {
-            purchaseFinalized(args, callbackContext);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.purchaseFinalized(args, callbackContext));
         } else if (action.equals("setSessionId")) {
-            setSessionId(args, callbackContext);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.setSessionId(args, callbackContext));
         } else if (action.equals("getSessionId")) {
-            getSessionId(callbackContext);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.getSessionId(callbackContext));
         } else if (action.equals("handleURLCallback")) {
-            handleURLCallback(callbackContext);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.handleURLCallback(callbackContext));
         } else if (action.equals("roktEvents")) {
-            roktEvents(args, callbackContext);
-            return true;
+            return executeRoktAction(callbackContext, delegate -> delegate.roktEvents(args, callbackContext));
         } else {
             return false;
         }
 
         callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.OK, args.getString(0)));
         return true;
+    }
+
+    private RoktPluginDelegate getRoktDelegate() {
+        if (!roktDelegateInitialized) {
+            roktDelegate = RoktKitAvailability.createDelegate();
+            roktDelegateInitialized = true;
+        }
+        return roktDelegate;
+    }
+
+    private interface RoktAction {
+        void run(RoktPluginDelegate delegate) throws JSONException;
+    }
+
+    private boolean executeRoktAction(CallbackContext callbackContext, RoktAction action) throws JSONException {
+        RoktPluginDelegate delegate = getRoktDelegate();
+        if (delegate == null) {
+            callbackContext.error(RoktKitAvailability.REQUIRED_MESSAGE);
+            return true;
+        }
+        action.run(delegate);
+        return true;
+    }
+
+    @Override
+    public void onDestroy() {
+        if (roktDelegate != null) {
+            roktDelegate.onDestroy();
+            roktDelegate = null;
+        }
+        super.onDestroy();
     }
 
     public void logEvent(final JSONArray args) throws JSONException {
@@ -442,211 +449,6 @@ public class MParticleCordovaPlugin extends CordovaPlugin {
 
             user.setConsentState(updatedState);
         }
-    }
-
-    private com.mparticle.kits.Rokt rokt() {
-        return MParticleRoktKt.getRokt(MParticle.getInstance());
-    }
-
-    private RoktConfig buildRoktConfig(final JSONObject configMap) throws JSONException {
-        if (configMap == null || configMap == JSONObject.NULL || configMap.length() == 0) {
-            return null;
-        }
-
-        final RoktConfig.Builder builder = new RoktConfig.Builder();
-        boolean hasConfig = false;
-
-        if (configMap.has("colorMode") && !configMap.isNull("colorMode")) {
-            final JSONObject colorModeObj = configMap.getJSONObject("colorMode");
-            if (colorModeObj.has("value") && !colorModeObj.isNull("value")) {
-                hasConfig = true;
-                builder.colorMode(RoktConfig.ColorMode.valueOf(colorModeObj.getString("value")));
-            }
-        }
-
-        if (configMap.has("cacheConfig") && !configMap.isNull("cacheConfig")) {
-            final JSONObject cacheConfigMap = configMap.getJSONObject("cacheConfig");
-            hasConfig = true;
-            final long cacheDuration = cacheConfigMap.has("cacheDurationInSeconds")
-                && !cacheConfigMap.isNull("cacheDurationInSeconds")
-                ? cacheConfigMap.getLong("cacheDurationInSeconds")
-                : 0L;
-            final JSONObject cacheAttributes = cacheConfigMap.has("cacheAttributes")
-                && !cacheConfigMap.isNull("cacheAttributes")
-                ? cacheConfigMap.getJSONObject("cacheAttributes")
-                : new JSONObject();
-            builder.cacheConfig(new CacheConfig(cacheDuration, ConvertStringMap(cacheAttributes)));
-        }
-
-        if (configMap.has("edgeToEdgeDisplay") && !configMap.isNull("edgeToEdgeDisplay")) {
-            hasConfig = true;
-            builder.edgeToEdgeDisplay(configMap.getBoolean("edgeToEdgeDisplay"));
-        }
-
-        return hasConfig ? builder.build() : null;
-    }
-
-    public void selectPlacements(final JSONArray args) throws JSONException {
-        final String identifier = args.getString(0);
-        final JSONObject attributesMap = args.getJSONObject(1);
-        final JSONObject configMap = args.isNull(2) ? null : args.getJSONObject(2);
-
-        final Map<String, String> attributes = ConvertStringMap(attributesMap);
-        final RoktConfig roktConfig = buildRoktConfig(configMap);
-
-        rokt().selectPlacements(
-            identifier,
-            attributes,
-            null,  // embeddedViews not used in Cordova
-            null,  // fontTypefaces not used in Cordova
-            roktConfig
-        );
-    }
-
-    public void selectShoppableAds(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-        final String identifier = args.getString(0);
-        final JSONObject attributesMap = args.getJSONObject(1);
-        final JSONObject configMap = args.isNull(2) ? null : args.getJSONObject(2);
-
-        final Map<String, String> attributes = ConvertStringMap(attributesMap);
-        final RoktConfig roktConfig = buildRoktConfig(configMap);
-
-        rokt().selectShoppableAds(identifier, attributes, roktConfig);
-        callbackContext.success();
-    }
-
-    public void purchaseFinalized(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-        final String identifier = args.getString(0);
-        final String catalogItemId = args.getString(1);
-        final boolean success = args.getBoolean(2);
-
-        rokt().purchaseFinalized(identifier, catalogItemId, success);
-        callbackContext.success();
-    }
-
-    public void setSessionId(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-        final String sessionId = args.getString(0);
-        if (sessionId != null && sessionId.length() > 0) {
-            rokt().setSessionId(sessionId);
-        }
-        callbackContext.success();
-    }
-
-    public void getSessionId(final CallbackContext callbackContext) {
-        String sessionId = rokt().getSessionId();
-        if (sessionId != null) {
-            callbackContext.success(sessionId);
-        } else {
-            callbackContext.success();
-        }
-    }
-
-    public void handleURLCallback(final CallbackContext callbackContext) {
-        // The Rokt Android SDK does not yet expose a redirect-URL callback;
-        // matches the no-op behaviour of the sibling RN and Flutter wrappers.
-        Logger.warning("handleURLCallback is not yet supported on Android");
-        callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.OK, false));
-    }
-
-    public void roktEvents(final JSONArray args, final CallbackContext callbackContext) throws JSONException {
-        final String identifier = args.getString(0);
-
-        // Tear down any prior subscription for this identifier so repeated calls don't
-        // stack duplicate collectors all delivering through different callbacks.
-        CoroutineScope previous = roktEventScopes.remove(identifier);
-        if (previous != null) {
-            CoroutineScopeKt.cancel(previous, (CancellationException) null);
-        }
-
-        Flow<RoktEvent> events = rokt().events(identifier);
-
-        Function2<RoktEvent, Continuation<? super Unit>, Object> onEach =
-            (event, continuation) -> {
-                JSONObject json = jsonFromRoktEvent(event);
-                if (json != null) {
-                    PluginResult result = new PluginResult(PluginResult.Status.OK, json);
-                    result.setKeepCallback(true);
-                    callbackContext.sendPluginResult(result);
-                }
-                return Unit.INSTANCE;
-            };
-
-        CoroutineScope scope = CoroutineScopeKt.CoroutineScope(Dispatchers.getDefault());
-        FlowKt.launchIn(FlowKt.onEach(events, onEach), scope);
-        roktEventScopes.put(identifier, scope);
-    }
-
-    @Override
-    public void onDestroy() {
-        // Cancel every active Rokt event subscription so collectors stop running and
-        // never deliver results through a callbackContext tied to a dead WebView.
-        for (CoroutineScope scope : roktEventScopes.values()) {
-            CoroutineScopeKt.cancel(scope, (CancellationException) null);
-        }
-        roktEventScopes.clear();
-        super.onDestroy();
-    }
-
-    private static JSONObject jsonFromRoktEvent(RoktEvent event) {
-        JSONObject json = new JSONObject();
-        try {
-            if (event instanceof RoktEvent.ShowLoadingIndicator) {
-                json.put("event", "ShowLoadingIndicator");
-            } else if (event instanceof RoktEvent.HideLoadingIndicator) {
-                json.put("event", "HideLoadingIndicator");
-            } else if (event instanceof RoktEvent.InitComplete) {
-                json.put("event", "InitComplete");
-                json.put("success", ((RoktEvent.InitComplete) event).getSuccess());
-            } else if (event instanceof RoktEvent.PlacementReady) {
-                json.put("event", "PlacementReady");
-                json.put("placementId", ((RoktEvent.PlacementReady) event).getIdentifier());
-            } else if (event instanceof RoktEvent.PlacementInteractive) {
-                json.put("event", "PlacementInteractive");
-                json.put("placementId", ((RoktEvent.PlacementInteractive) event).getIdentifier());
-            } else if (event instanceof RoktEvent.PlacementClosed) {
-                json.put("event", "PlacementClosed");
-                json.put("placementId", ((RoktEvent.PlacementClosed) event).getIdentifier());
-            } else if (event instanceof RoktEvent.PlacementCompleted) {
-                json.put("event", "PlacementCompleted");
-                json.put("placementId", ((RoktEvent.PlacementCompleted) event).getIdentifier());
-            } else if (event instanceof RoktEvent.PlacementFailure) {
-                json.put("event", "PlacementFailure");
-                String placementId = ((RoktEvent.PlacementFailure) event).getIdentifier();
-                json.put("placementId", placementId != null ? placementId : JSONObject.NULL);
-            } else if (event instanceof RoktEvent.OfferEngagement) {
-                json.put("event", "OfferEngagement");
-                json.put("placementId", ((RoktEvent.OfferEngagement) event).getIdentifier());
-            } else if (event instanceof RoktEvent.PositiveEngagement) {
-                json.put("event", "PositiveEngagement");
-                json.put("placementId", ((RoktEvent.PositiveEngagement) event).getIdentifier());
-            } else if (event instanceof RoktEvent.FirstPositiveEngagement) {
-                json.put("event", "FirstPositiveEngagement");
-                json.put("placementId", ((RoktEvent.FirstPositiveEngagement) event).getIdentifier());
-            } else if (event instanceof RoktEvent.OpenUrl) {
-                RoktEvent.OpenUrl openUrl = (RoktEvent.OpenUrl) event;
-                json.put("event", "OpenUrl");
-                json.put("placementId", openUrl.getIdentifier());
-                json.put("url", openUrl.getUrl());
-            } else if (event instanceof RoktEvent.CartItemInstantPurchase) {
-                RoktEvent.CartItemInstantPurchase purchase = (RoktEvent.CartItemInstantPurchase) event;
-                json.put("event", "CartItemInstantPurchase");
-                json.put("placementId", purchase.getIdentifier());
-                json.put("cartItemId", purchase.getCartItemId());
-                json.put("catalogItemId", purchase.getCatalogItemId());
-                json.put("currency", purchase.getCurrency());
-                json.put("description", purchase.getDescription());
-                json.put("linkedProductId", purchase.getLinkedProductId());
-                json.put("totalPrice", purchase.getTotalPrice());
-                json.put("quantity", purchase.getQuantity());
-                json.put("unitPrice", purchase.getUnitPrice());
-            } else {
-                json.put("event", event.getClass().getSimpleName());
-            }
-        } catch (JSONException e) {
-            Logger.warning(e, "Failed to serialize RoktEvent");
-            return null;
-        }
-        return json;
     }
 
     private static IdentityApiRequest ConvertIdentityAPIRequest(JSONObject map) throws JSONException {
